requirements.txt
openai-whisper
torch
torchaudio
numpy
soundfile
gTTS
googletrans==4.0.0-rc1
